関数signup内のfirebaseに関する記述をaction経由に変更
